backend/pyproject.toml
[project]
name = "seamless-api"
version = "0.1.0"
description = "FastAPI translation service powered by SeamlessM4T"
authors = [{ name = "Hydra Dynamix", email = "bakobiibizo@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydub>=0.25.1",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    # SeamlessM4T from Facebook Research
    "seamless_communication @ git+https://github.com/facebookresearch/seamless_communication.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
seamless-api = "seamless_api.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["seamless_api"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
